*** empty log message ***

+++ utilities/composite.c
@@ -317,7 +317,7 @@ static unsigned int CompositeImageList(ImageInfo *image_info,Image **image,
                 }
                 case NorthEastGravity:
                 {
-                  geometry.x+=geometry.width-composite_image->columns;
+                  geometry.x+=(long) geometry.width-composite_image->columns;
                   break;
                 }
                 case WestGravity:
@@ -342,21 +342,21 @@ static unsigned int CompositeImageList(ImageInfo *image_info,Image **image,
                 }
                 case SouthWestGravity:
                 {
-                  geometry.y+=geometry.height;
-                  geometry.y+=geometry.height-composite_image->rows;
+                  geometry.y+=(long) geometry.height;
+                  geometry.y+=(long) geometry.height-composite_image->rows;
                   break;
                 }
                 case SouthGravity:
                 {
                   geometry.x+=(long)
                     (geometry.width/2-composite_image->columns/2);
-                  geometry.y+=geometry.height-composite_image->rows;
+                  geometry.y+=(long) geometry.height-composite_image->rows;
                   break;
                 }
                 case SouthEastGravity:
                 {
-                  geometry.x+=geometry.width-composite_image->columns;
-                  geometry.y+=geometry.height-composite_image->rows;
+                  geometry.x+=(long) geometry.width-composite_image->columns;
+                  geometry.y+=(long) geometry.height-composite_image->rows;
                   break;
                 }
               }
